delete test mode evaluator
