+++ chainercv/__init__.py
@@ -3,6 +3,7 @@ import extensions  # NOQA
 import functions  # NOQA
 import models  # NOQA
 import testing  # NOQA
-import training  # NOQA
+import transforms
 import utils  # NOQA
 import wrappers  # NOQA
+
