--- chainercv/training/test_mode_evaluator.py
@@ -1,11 +1,0 @@
-from chainer.training import extensions
-
-
-class TestModeEvaluator(extensions.Evaluator):
-
-    def evaluate(self):
-        model = self.get_target('main')
-        model.train = False
-        ret = super(TestModeEvaluator, self).evaluate()
-        model.train = True
-        return ret
